UserManagement/UserMgmt/requirements.txt
aniso8601==8.0.0
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
Flask==1.1.2
Flask-JWT-Extended==3.24.1
flask-mongoengine==0.9.5
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
kubernetes==12.0.1
MarkupSafe==1.1.1
mongoengine==0.19.1
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
pytz==2019.3
six==1.14.0
Werkzeug==1.0.1
WTForms==2.3.1
uWSGI==2.0.18
pycryptodome==3.10.1
